--- .npmignore
@@ -1,4 +1,0 @@
-test
-.github
-.eslintrc
-karma.conf.js
