+++ package.json
@@ -2,7 +2,10 @@
   "name": "@bpmn-io/execution-platform",
   "version": "0.2.0",
   "description": "Add execution platform meta-data to BPMN and DMN diagrams",
-  "main": "index.js",
+  "main": "src/index.js",
+  "files": [
+    "src"
+  ],
   "scripts": {
     "all": "run-s lint test",
     "lint": "eslint .",
