+++ src/cmd/UpdateExecutionPlatformCommand.js
